backend/services/transaction-service/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
Pillow==10.2.0
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.1
razorpay==1.4.2
